Fixed DMA error, implemented image size detection

+++ tracker/software/types.h
@@ -112,8 +112,8 @@ typedef struct {
 	resolution_t res;		// Camera resolution
 	uint8_t quality;		// JPEG quality
 	uint8_t *ram_buffer;	// Camera Buffer
-	uint16_t ram_size;		// Size of buffer
-	uint16_t size_sampled;	// Actual image data size (do not set in config)
+	uint32_t ram_size;		// Size of buffer
+	uint32_t size_sampled;	// Actual image data size (do not set in config)
 	bool redundantTx;		// Redundand packet transmission (APRS only)
 } ssdv_conf_t;
